Conflicts:
	BizHawk.Client.Common/BizHawk.Client.Common.csproj
	BizHawk.Client.EmuHawk/DisplayManager/DisplayManager.cs
	BizHawk.Common/InstanceDll.cs
	BizHawk.Emulation.Cores/Consoles/Nintendo/SNES/LibsnesCore.cs
	vbanext/instance.cpp

+++ BizHawk.Client.EmuHawk/CustomControls/VirtualListView.cs
@@ -19,8 +19,8 @@ namespace BizHawk.Client.EmuHawk
 	[StructLayout(LayoutKind.Sequential)]
 	internal struct NmHdr 
 	{
-		public int HwndFrom;
-		public int IdFrom;
+		public IntPtr HwndFrom;
+		public IntPtr IdFrom;
 		public int Code;
 	}
